Fixed client prediction not reconciling
There was a beginner's mistake with the server not always sending the world state if the position did not change. This mean the client would predict that movement would happen, but never get a correction state packet from the server saying otherwise unless the server acknowledged movement of some kind. Eg - up against a wall, press direction of the wall, client doesnt have the wall enabled and it would be considered valid. This would allow cheating.